journal: move LeaderBasedJournal to seperate mod
