+++ src/journal/src/remote/mod.rs
@@ -12,4 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod leader_based;
 mod shared;
